renamed AppointmentDetail to AppointmentCalendarDetail
